Refactor related newer hooks together into the same folder.
