+++ apps/vor/frontend/src/api/classes/useGetClass.ts
@@ -1,5 +1,5 @@
 import { QueryResult, useQuery } from "react-query"
-import { fetchApi } from "./fetchApi"
+import { fetchApi } from "../fetchApi"
 
 interface Class {
   id: string
