+++ apps/vor/frontend/src/components/PageChooseSchool/PageChooseSchool.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react"
 import { navigate } from "gatsby"
 import Flex from "../Flex/Flex"
 import Box from "../Box/Box"
-import { useGetSchools } from "../../api/useGetSchools"
+import { useGetSchools } from "../../api/schools/useGetSchools"
 import Card from "../Card/Card"
 import { Typography } from "../Typography/Typography"
 import Icon from "../Icon/Icon"
