+++ apps/vor/frontend/src/components/PageEditClass/PageEditClass.tsx
@@ -10,9 +10,9 @@ import Flex from "../Flex/Flex"
 import Chip from "../Chip/Chip"
 import Button from "../Button/Button"
 import { WEEKDAYS } from "../PageNewClass/PageNewClass"
-import useGetClass from "../../api/useGetClass"
+import useGetClass from "../../api/classes/useGetClass"
 import LoadingPlaceholder from "../LoadingPlaceholder/LoadingPlaceholder"
-import usePatchClass from "../../api/usePatchClass"
+import usePatchClass from "../../api/classes/usePatchClass"
 import LoadingIndicator from "../LoadingIndicator/LoadingIndicator"
 import DeleteClassDialog from "../DeleteClassDialog/DeleteClassDialog"
 import ErrorMessage from "../ErrorMessage/ErrorMessage"
