+++ apps/vor/frontend/src/components/PageEditGuardians/PageEditGuardians.tsx
@@ -3,7 +3,7 @@ import BackNavigation from "../BackNavigation/BackNavigation"
 import {
   Guardians,
   useGetSchoolGuardians,
-} from "../../api/useGetSchoolGuardians"
+} from "../../api/guardians/useGetSchoolGuardians"
 import { Student, useGetStudent } from "../../api/useGetStudent"
 import { Typography } from "../Typography/Typography"
 import { Card } from "../Card/Card"
@@ -16,10 +16,10 @@ import { ReactComponent as LinkIcon } from "../../icons/link.svg"
 import { ReactComponent as RemoveIcon } from "../../icons/close.svg"
 import { Flex } from "../Flex/Flex"
 import GuardianRelationshipPickerDialog from "../GuardianRelationshipPickerDialog/GuardianRelationshipPickerDialog"
-import { usePostGuardianRelation } from "../../api/usePostGuardianRelation"
+import { usePostGuardianRelation } from "../../api/guardians/usePostGuardianRelation"
 import { Link } from "../Link/Link"
 import { Button } from "../Button/Button"
-import { useDeleteGuardianRelation } from "../../api/useDeleteGuardianRelation"
+import { useDeleteGuardianRelation } from "../../api/guardians/useDeleteGuardianRelation"
 import Dialog from "../Dialog/Dialog"
 import LoadingIndicator from "../LoadingIndicator/LoadingIndicator"
 import GuardianRelationshipPill from "../GuardianRelationshipPill/GuardianRelationshipPill"
