+++ apps/vor/frontend/src/components/PageNewGuardian/PageNewGuardian.tsx
@@ -8,7 +8,7 @@ import { navigate } from "../Link/Link"
 import { EDIT_GUARDIANS_URL } from "../../routes"
 import LoadingIndicator from "../LoadingIndicator/LoadingIndicator"
 import { Box } from "../Box/Box"
-import { usePostNewGuardian } from "../../api/usePostNewGuardian"
+import { usePostNewGuardian } from "../../api/guardians/usePostNewGuardian"
 import BackNavigation from "../BackNavigation/BackNavigation"
 import { Typography } from "../Typography/Typography"
 import { useGetStudent } from "../../api/useGetStudent"
