+++ apps/vor/frontend/src/components/PageNewPlan/PageNewPlan.tsx
@@ -7,7 +7,7 @@ import BackNavigation from "../BackNavigation/BackNavigation"
 import { ALL_PLANS_URL } from "../../routes"
 import { Typography } from "../Typography/Typography"
 import Button from "../Button/Button"
-import useGetSchoolClasses from "../../api/useGetSchoolClasses"
+import useGetSchoolClasses from "../../api/classes/useGetSchoolClasses"
 import Chip from "../Chip/Chip"
 import { Flex } from "../Flex/Flex"
 import TextArea from "../TextArea/TextArea"
