+++ apps/vor/frontend/src/components/PageNewStudent/PageNewStudent.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState } from "react"
 import { useImmer } from "use-immer"
 import { Link, navigate } from "../Link/Link"
-import { useGetGuardian } from "../../api/useGetGuardian"
+import { useGetGuardian } from "../../api/guardians/useGetGuardian"
 import Box from "../Box/Box"
 import BackNavigation from "../BackNavigation/BackNavigation"
 import Input from "../Input/Input"
@@ -10,7 +10,7 @@ import DateInput from "../DateInput/DateInput"
 import TextArea from "../TextArea/TextArea"
 import { Typography } from "../Typography/Typography"
 import Select from "../Select/Select"
-import useGetSchoolClasses from "../../api/useGetSchoolClasses"
+import useGetSchoolClasses from "../../api/classes/useGetSchoolClasses"
 import Chip from "../Chip/Chip"
 import { Flex } from "../Flex/Flex"
 import LoadingPlaceholder from "../LoadingPlaceholder/LoadingPlaceholder"
